Use blockscope {} to correct some "goto crosses init"
The compiler warns you if you use a goto to cross a variable init,
because it means you can't have it initialized on the goto's branch.

It's possible to avoid with a scope, but that can look kind of random.
`#define blockscope` was added to help with that, and makes these
situations legible without needing a warning.  Creative 2-space
outdenting makes it not too invasive.