Enforce VAL_ARRAY_AT(), VAL_STRING_AT() in bounds
Systemically, Rebol has never had a thorough answer to what pieces of
code are responsible for ensuring that an array index is "stale" and
out of bounds.  Running the Red tests push on edge cases that show
that it's easy to do, and some level of the system has to take
responsibility or there will be crashes.

This puts a layer of safety in at the VAL_XXX_AT() routines dealing
with indexes, leaving the lower level XXX_AT() unchecked and assuming
that code there knows what it's doing.

These changes should not affect existing code (unless it was code that
could have crashed, that will now get an error instead.)