Set current "directory" (WHAT-DIR) from URL in %load-r3.js
Scripts have to have an idea of what the "current directory is" when
they are running.  If a resource is requested as a FILE! (as opposed
to an absolute URL!) it is fetched by path relative to that.  What
makes the most sense as the starting "directory" on the web is what's
visible in the URL bar.  Then, executing URLs can change the "current"
notion to whatever scripts you DO, while they are running.

Also updates some comments.