Add `?local` feature to load-r3.js, reorganization/cleanup
When developing locally, it's convenient to be able to use the same
%load-r3.js code (either deployed or local) and not have to edit it
by hand to tell it to use local files.  However, cueing this from
`localhost` is not the best idea, because a curious user might want
to develop a page locally and view it while fetching the library
remotely.  It makes more sense to be convenient for them and make those
doing their own builds of the library do a little more work.

This makes it easier by letting you say ?local in the URL.  This means
it will look to fetch libr3.js out of the 0.16.1 and 0.16.2 of the
page that is invoking the library.

Also does some tidying of the code, removing old aliases for certain
routines, using `let` instead of `var` where applicable, etc...