Do KEEPALIVE and EMTERPRETER_BLACKLIST with JSON files
Previously the emscripten keepalive was done with macros that had to
be on the functions themselves.  Since "rebol.h" does not actually want
to include "emscripten.h", this meant faking the annotation's
definition.  There is an alternative of providing a list to emcc on
the command line, which can also be done with a .json file.  This
changes to a programmatic approach where the list is built from the
APIs as a root set and set as JSON.

A similar JSON-file-based capability is available for the emterpreter
blacklist.  This uses that facility, and adds more API implementation
routines (e.g. RL_rebInteger()) to the blacklist.  This doesn't mean
they can be called yet in an awaiter, as emscripten's cwrap prohibits
it...so either cwrap would need changing, or we'd not use it.