Workaround for "tolerant" cwrap and ccall
A facility was added to make the emterpreter tolerant of use of a
function on the EMTERPRETER_BLACKLIST (e.g. pure WASM) during an
emterpreter_sleep_with_yield():

https://stackoverflow.com/q/51204703/

But that only speaks to the tolerances of the bytecode interpreter
itself.  So although you can call the underscore-prefixed version of
the function, the `ccall` function (used in the implementation of
cwrap) did not take this into account.  A cwrap'd blacklisted function
does not let you call it.

As a short-term hack, this doesn't ask for the inclusion of `cwrap`
or `ccall` in EXTRA_EXPORTED_RUNTIME_METHODS.  Instead it holds
copies of those functions minus the assert.  A longer-term solution
is probably to customize the calling procedure, since there are so
few entry points that it wouldn't be too hard (and we're already
implementing it for variadic functions anyway).