Transcode bootstrap fixes
Though it was believed that the bootstrap had been tested for the
transcode changes, it appears that some mistake must have been made
and not been using a bootstrap executable--because there were some
bugs in it that should have prevented bootstrap if it had actually
been tested in the way that was believed.

This addresses those issues.

(Also converts some stray line endings from Windows to Unix style.)