Fix promise result handling in JavaScript extension
An attempt at optimization used the CELL_FLAG_MARKED bit to indicate
the difference between a promise returning an actual ERROR! vs.
failing with an ERROR!.  It turned out this didn't work, because the
cell was moved by the rescue operation and lost the mark bit.  Because
the mark bit is a very nuanced thing at this point, it's probably
better to forget about the optimization and use a separate boolean...
and revisit the question later.