Make REBSTR its own subtype
Handling of UTF-8 strings is very different from how other series are
handled, so it's best to make them type incompatible at compile-time.
This uses the same technique as REBARR and other types, providing a
helper function STR() to manually annotate a cast from a REBSER to a
REBSTR.

Having a separate type suggests that there can be a STR_LEN() method
which knows the codepoint length should be fetched from the MISC()
field and doesn't need a test.  Most calls in the system should know
whether they have a string or not, and hence use the right length
function.

Also gets rid of the LENGTH OF support in ANY-WORD!.  It is likely
better to have ANY-WORD! act as atomics.