Re-unify Modify_String() and Modify_Binary()
UTF-8 made the code for Modify_String() more complex than that for
Modify_Binary(), because the size of codepoints was not fixed.  So
they were split into two separate routines.

However, the ultimate goal was to make it possible to alias between
UTF-8 strings and binary series...but keeping the invariant that the
aliased binaries must always point to valid UTF-8 data.  This means it
would be possible to be appending two binaries that were actually
two aliased UTF-8 strings.  Hence the complexity of binary modificaiton
can be necessarily as complex as string modification.

This pushes that unification forward a bit, making the AS operation
actually do the aliasing (which was stubbed out via copying before).