Allow adding BINARY! content to strings (is validated)
This allows non-previously-string-aliased BINARY! data to be appended
to strings, if it passes a UTF-8 validation.

Also fixes a rendering issue in MOLD where a BINARY! that was an alias
of a string series was getting the string's length, because it was
using the generic "SER_LEN()" which had been detecting string series
and giving the length from that.  This is another good reason (besides
just performance) to deprecate SER_LEN() and have clients know what
they are asking about.