First step at making refinement args their own parameters
This is a work-in-progress which does a tweak to copy a refinement
argument's value to the refinement slot during the type check.  This is
a poor-man's simulation of an actual evaluator overhaul which lets the
refinement itself carry the argument.

The internal routines are modified by way of having the parameter
generation macros modified to not generate ARG() accessors for what
were formerly "refinement arguments".  At this stage, the only external
commands that are modified to the new convention are those enough to
get it to boot--as a proof of concept.

The only examples of multiple parameters to a refinement that were
encountered were the AES and RC4 cryptography routines.  They were
not examples of a good use of such refinements, as they actually seemed
to try and cram two routines into one--possibly because the extension
they were in made it difficult to make new commands.

While much more work is needed before this will be ready, it is a very
promising start--which is very clarifying.  The C source benefits
significantly from not having the "invented" names for refinement
arguments to deal with.