Potentially manage bindings in UNWIND
It is possible to be asked to run an UNWIND when an output cell does
not have stack lifetime (e.g. if the cell is being produced for the
API).  So even though a throw to a function would be caught at that
level, the asserts are checking to make sure a non-reified frame
pointer doesn't get put into a cell that could outlive it.

This needs more review, as the avoidance of reification is speculative
and hasn't been fully realized yet.  However, this at least stops
the assert.

Also changes the /WITH form to be like RETURN so that the optional
return value just comes inline as an `<end>`-able argument.

Makes separate test file %functions/unwind.reb to collect tests for
UNWIND specifically.