Make COMPOSE honor newline status of GROUP!
This makes a COMPOSE cue the newline status of inserted material from
the GROUP! in the source.  So in the following two examples, the
newline status of `thing` doesn't matter, it will be on the same line
in the `<a>` case and on its own line in the `<b>` case regardless.

    compose [<a> (thing)]

    compose [
        <b>
        (thing)
    ]

The spliced case is less obvious, but it is done to use the GROUP!'s
status for the first element and then the content's flags for the
remainder of the elements.