Update %redbol.reb to new parameter conventions, other fixes
This brings the fledgling %redbol.reb up to date (enough that it can
be used to generate the blog entries on hostilefork.com, despite that
generator being Rebol2-style code).

It incorporates a bridge from the old refinement argument convention
to the new one.  It also brings the non-const forms of function in,
so that one can modify strings and arrays in the body of the emulated
forms of functions.