Handle multi-line binaries pasted in console input
There are two phases in which an error on a binary can occur, one is
scanning for the end brace (same code as for strings) and the next
phase checks that it has valid hex or base64/etc. for the binary.

These were both reporting the same error of an invalid binary.  But
converting the first error to "missing }" allows it to be recognized
by the same code that allows the console to continue gathering lines
of input.

It does mean that you can be entering multi-line binaries with bad
hex digits and it will keep accepting lines in the console and not
tell you the binary is bad until the end.  But considering that it is
letting you paste multi-line binaries, it's fine.