Changes for Redbol COMPOSE to support VOID! vaporizing
    rebol2>> type? either true [] []
    == unset!

    rebol2>> compose [(either true [] [])]
    == []

Ren-C fundamentally wants taken branches that do not produce content
to be VOID! (so they can be used with ELSE) but doesn't want voids
to vaporize in COMPOSE.  The way to work around this in the emulation
layer for Rebol2 is with a "predicate" function that is run on the
composed slot.

Implementing this feature motivated the creation of a DEVOID operation
that would turn voids into nulls, and to fix the predicate feature
(which hadn't really been tested very much) for tolerance of when
the predicate produce a NULL.  It also corrects an assert caused by
fulfilling arguments with void, where the caller of an optimization is
spposed to ensure it wasn't void before the call...so that callers who
already knew it wasn't don't have to pay for the check in the routine.