Catch cases of WRITE without /BINARY in Redbol
Rebol2 apparently would inject CR LF sequences on LF bytes when using
WRITE with a BINARY!.  This is confusing.  This adds a /BINARY
refinement to the emulation for WRITE, along with the other Rebol2
refinements to error on as not implemented.  It then refuses to write
a BINARY! unless /BINARY is used.