Add error message for missing data descriptor
The zip file format is able to write its data to a stream, where it
can't go back and patch up header information after-the-fact to add
sizes.  In these cases, it puts zero in the header slots for the
compressed size, uncompressed size, and CRC.

The Rebol Unzip code wasn't written to handle this situation.  This
adds an error upon recognizing the case.