Allow override of greeting via console skin
There was an apparently unfinished feature of the console skin giving
a customized greeting.  It also wasn't possible to override the
console skin filename, or to create a skin object dynamically without
a file.  This adds a couple of changes to permit it, aiming at letting
the web console customize its greeting.