Move definitions from %s-unicode.c file to %sys-char.h
The R3-Alpha codebase had taken a header and c file that was from the
Unicode group, and hacked it up somewhat for usage.  There was a fair
amount of unused code in it related to UTF16 conversions.

Much of that code was commented out.  Given UTF-8's pervasiveness and
Ren-C's choice to move to "UTF-8 Everywhere" for the internal
representation of strings, it's not likely to be needed.  The small
amount of code that is needed is best to have in the header files so
the optimizer can make decisions about how to inline it (partially
or fully), since UTF-8 scanning is fairly critical.

Also, the file had its own definitions for Boolean and types for
UTF8, UTF16, and UTF32...which were superfluous and causing warnings
(a build for HaikuOS was triggering those warnings, and motivated
cleaning this up).

Also deletes an unused %t-vector.c file.  The actual vector code now
is in an extension.