Redbol fallback to Latin1 if UTF-8 decode fails
Files that use extended ASCII as Latin1 are unlikely to be valid UTF-8.
Hence failing to load as UTF-8 in Redbol can be a signal to fall back
and load as Latin1...thus being compatible for LOAD with Rebol2.

(Includes patches to try and build on HaikuOS with no warnings.)