Make Generic Dispatch return R_UNHANDLED on failure
Reporting a good error message on generic action dispatch had involved
passing the REB_XXX type to the error.  This becomes more complex with
the introduction of extension types.

The Generic_Dispatcher() has all the information that it needs to
implement this error centrally.  And in the long tail of that handler,
whatever "Generic" dispatch mechanism there is will ultimately be
responsible for knowing about whether the methods are registered...not
a hard-coded switch() statement.

Knowing that day is coming, it's cleaner to move the error into the
dispatcher so that the extension type errors don't have to get more
complicated