Make LIBRARY! use extension type mechanism
In order to reclaim a scarce Builtin-Type, this makes the LIBRARY!
datatype use the extension type mechanism.  However, it is not moved
to an extension...it just follows the extension type cell format.

It's not much of a sacrifice for LIBRARY!--as it was only using one
of its 3 cell fields.  So sacrificing the EXTRA to be the datatype is
not a problem.