Return ISSUE! to being an ANY-STRING!
The R3-Alpha change for ISSUE! to being an ANY-WORD! was motivated by
the desire to get the symbol-optimization benefit of WORD! in an inert
type.  Word types were more scarce than strings, so the tradeoff was
considered to be a useful one.

Carl had received feedback that this was controversial (and I saw
firsthand how much Rebol2 code this broke, e.g. when trying to port
Red to R3-Alpha, back when that seemed like a good idea).  Though no
one really seemed much in favor of it, Red went with it.

Yet it's fundamentally flawed in terms of the existing usages, which
very clearly used things that are not legal words--like #123.  This
means issues were creating bindings to illegal words left and right:

    >> #123
    == #123

    >> words-of bind? #123
    == [system 123 words-of bind?]

The common practice of using ISSUE! for numbers shows a clear non-fit.

Ren-C introduces @word as a new ANY-WORD! type, which is also inert.
This makes it pick up the slack as the new favored form of word to use
when speaking about an item in a fixed enumeration, which you do not
expect people to be evaluating--hence having it inert after any number
of COMPOSE or REDUCE operations is an asset.

With the addition of the new word type, the responsibilities that
ISSUE! had in the system for being a "weird inert word type" can be
taken over, and ISSUE! is returned to being a string.