Extract File, Directory, Serial ports to extensions
This moves several bodies of port code that cannot be used in the
emscripten build into individually selectable extensions.  The result
of removing them takes about 30k out of the optimized download size.

Accomplishing this required removing the static table by which "Rebol
Devices" were registered, replacing it with a linked list.  To be able
to factor code out so that the core could build without the extension,
some callbacks had to be shuffled in terms of the "host" API (which is
being ultimately targeted for deletion).