Move Dev_Event, WAIT, WAKE-UP into the Event extension
Although the port and datatype dispatching for EVENT! had been moved
into an optional extension not included in the Emscripten build, the
"device" had not been moved.  The last few commits have made it
possible to move devices out of the core, though this was a bit tricky
because the WAIT command depended on it.

Removing the WAIT command from core and having it supplied by an
extension makes sense, since how waiting is done depends very much on
the host environment (as opposed to the language interpreter).  It
means that the JavaScript build has to provide its own implementation,
but it already meant that anyway.