Move POSIX Signal Port to an Extension
The POSIX Signal Port does not have a lot of documentation, and it was
only enabled on Linux.  This changes its mode of enablement to be
via an extension that is requested or not.  However, the compiler
definition HAS_POSIX_SIGNAL is still used to determine what kind of
jump buffer longjmp/setjmp use.