NOW moved to Time Extension
The idea of making time itself an extension is a bit odd, as the C
standard includes functions for it.  However, R3-Alpha seems to have
wanted more flexibility (e.g. nanosecond resolution) and so it was
part of the "host kit", meaning there were APIs like OS_GET_TIME().

Ren-C's goal is to eliminate the IOCTL-like table of functions that
were host kit, in favor of a more granular approach with extensions.
Getting rid of it requires killing off OS_GET_TIME() one way or
another.  This takes the "easiest" route by moving NOW to an extension,
and putting OS_DELTA_TIME() into the event extension (which is the
only place it was used).

As a consequence of this move, it's no longer possible to use the
time extension to time the boot process.  That would have to be done
some other way.  It also means that the JavaScript module has to
either use the extension or reimplement it...and it makes more sense
to reimplement the NOW function using a JS-NATIVE, because that cuts
out the middleman of the Emscripten emulation of POSIX time.