Eliminate posix host-process.c, host-event.c
The POSIX %host-process.c file had turned mostly into the Process
extension implementing CALL (among other things).  Only one routine
remained, which was OS_Reap_Process().  This commit folds that
routine in with the Event extension--which is the only place it was
called.

Also this gets rid of the `host-event.c` file for Linux, as it was
never built in Ren-C...and was an X-Windows-based stray file that
was kept around due to keeping other EVENT!-oriented code.  To the
extent this code is relevant in 2019, the relevant copy would be in
Atronix's R3/View codebase...and that should be what is updated to
the new conventions.