Move TO-LOCAL-FILE, WHAT-DIR, etc. to File extension
This does some more shuffling with the ultimate goal of being able to
eliminate the OS_XXX APIs.  It is a bit rough in how it accomplishes
this, given that modules aren't making it easy to link up the formerly
core definitions to live in LIB.  However, it pushes logic like that
which translates paths into the file extension.

It also takes the code for finding the currently running executable
and puts that into the Filesystem extension as well, as a native.