Move LIBRARY! datatype to an extension
This moves the DLL support code to an extension.  While it is not
currently tested (due to DLL/FFI features being on the back burner at
this precise moment), it should be easier to work with and test now
that it is a clearly separated component.  And the code would not be
of use in the JavaScript build, so being an extension lets it be
excluded and reduces the size of the download.