Eliminate %reb-host.h, %host-table.c, %make-os-ext.r...
This commit is the culmination of the previous extractions of all
OS-specific features into extensions.  It eliminates the "host kit"
entirely, removing the need for several obfuscating steps in the
build process.

Philosophically, what this gets rid of is the idea of running on an
"abstract OS" which offers some table of entry points via C---but
that does not know anything about Rebol data types.  This abstract OS
could thus ostensibly be used for purposes other than Rebol.  Doing
such a thing well would be very difficult--and the host kit did not
do it well.

Using extensions which communicate with the system (and each other)
through Rebol values and natives has much more benefit.  It also has
a better granularity than one large "host table" of services, because
one only needs implement the extensions for an OS where it is needed.
One can choose either to expand on the build options for an existing
extension, or to start a new extension entirely (the JavaScript build
is currently not using most of the extensions, e.g. for I/O)