Add DO onto %load-r3.js execution of SCRIPT tags
The execution of scripts was using `rebElide()`, which does not go
through DO or LOAD implicitly.  As a result, plain evaluation will not
give effects for modules.  So:

   reb.Elide("Rebol [Type: Module Options: [isolate]] ...")

Here REBOL is an OBJECT!, which it evaluates to and throws away.  Then
the next element is a BLOCK! which evaluates to itself and is also
discarded.  That means no effect will be seen on the ensuing code.

To work around this today, clients who believe the code they are
executing is the source code of a script have to use DO / LOAD as
an explicit part of the execution.