Do_Any_Array_At_Throws() when array and output are same
For efficiency, Do_Any_Array_At_Throws() allows the target cell to
which output is written to be the same as the cell holding the array.
This is because everything needed from the input cell is extracted
into the frame feed.

It may not be the best idea in the world, but it does save on having
to find or create another GC-safe cell in a lot of occasions.  However,
a bug had been introduced where the feed was being initialized
from the array after the output was initialized to void.