Add API-TRANSIENT (support for JS-EVAL/JS-DO dialect)
This adds a feature which allows the creation of a transient API
handle (e.g. like one that rebR() has been called upon) from any
Rebol value, returned as an INTEGER! handle.

What it's for is the implementation of cleaner substitutions in
the inline processing of values when calling JavaScript eval().