DO of a module-based script returns VOID!
Because it was based on IMPORT that only returns a MODULE!, DO of a
module-based script was previously returning that module.  However,
this makes wrapping scripts as modules for features like isolation
difficult and incompatible.

This is a near term change that returns void for modules, so that if
all a module does is print messages those messages can be seen instead
of the result.  Ultimately it should return the result from the module,
but since it is based on IMPORT, and import only returns a MODULE!,
that's not currently possible.