Module pattern, load_r3=>reb.Startup taking config parameter
This changes `load_r3` from a Promise, to become reb.Startup(), a
function which takes an (optional) config object and returns a promise.

The config object allows specifying hooks for log(), warn(), error(),
and info().  The hooks default to the console functions of the same
name, if not overridden.

This replaces a previous strategy where %load-r3.js encoded specific
knowledge of a div named `replpad`, and would rewrite the console.xxx()
functions to direct to that.  Not permanently overriding the console
is important, because it was difficult to debug replpad output
functions if `console.log` could not be used without causing recurison.
But more generally, it's not obvious that `console.log` the JS function
should always be writing to the replpad's output.

In addition to these changes, the contents of load-r3.js are now all
wrapped up in the `reb.Startup()` function.  That means the internal
variables do not leak into the global namespace of the client.  The
only leakages are `Module` (because Emscripten requires it to be set
up in order to work) and `reb` (which ultimately holds the API, but
starts out holding only the Startup() function).