Unify JS-NATIVE and JS-AWAITER return result methods
This simplifies the code a slight bit, by making the JS function
objects carry an `is_awaiter` property so they know if they are one,
and hence execute the RunNative differently.

But in both cases, they use the same protocol of signaling an accept
or a reject.  This should provide a way to process errors out of a
plain JS-NATIVE with parity with how JS-AWAITERs are handled.

Also adds M or P character marker on debug trace lines to show whether
the log line comes from the "Main" thread or the "Promise" thread.