Keep main thread free during Rebol code on promise
In the first try at the design, the concept was that if the Rebol
interpreter was running on the worker thread that the main thread
should be blocked until some operation needed to unblock it (e.g. a
request for I/O).  This would avoid errors related to trying to use
the libRebol API from the GUI thread while the worker was running--
which is not safe.

However, locking the main thread is limiting--and so protection should
be done some other way.  Keeping the UI responsive allows for things
like cancellation signals to be received during a tight loop which has
no specific requests for I/O.  It was also very annoying because the
browser's GUI thread would frequently be locked during crashes, and
the page was difficult to close and reload.

This is a significant rethinking of how to accomplish synchronization,
which leverages the use of MAIN_THREAD_EM_ASM().  This allows a much
cleaner vehicle for reuse of code that needs to do some execution on
the main thread...removing the need for a hacky method of moving the
JS-NATIVE registration to the main thread.