Extract message from error for throw/reject()
There was only one case for error handling that was handled at all
previously.  It tried a dicey mechanism that's built on throw() which
doesn't allow it to be caught by Rebol's TRAP.  But it tunnels a
JavaScript error upwards to (theoretically) be caught by JavaScript
that called the Rebol.

Though just an experiment, it did stop the console from crashing for a
limited set of cases.  But it had an enigmatic message of "JavaScript
error not in rebPromise()".  This is because there was no clear way
for the worker to ask the main thread to stringify the error when it
was blocked.

With the main thread unblocked, it's possible to rig something up.
This does that, as well as further consolidate the resolve/reject
forms for ordinary natives and awaiters on using thes same handlers.