Only use bitfield header puns in Linux Builds
In order to make it easier in the debug watchlist to see which bits in
a Reb_Header were set, little endian builds w/GCC used "type punning"
to alias the 32-bits worth of flags as bitfield structures.

It turns out that Mingw64 on Cygwin says it is gcc, but does not have
the same bitfield packing outcome...making larger than 64-bit size
structures for what look like they "should" pack neatly into 32-bits.

This isn't too much of a surprise since bitfields are not very well
specified in the C standard.  In any case, the feature is probably not
worth it and tends to get out of date...but rather than delete it
outright this only puts it in the debug linux build.