Check return types for JavaScript natives
Due to the JavaScript extension not being a debug build, the code which
asserted type checking had been done wasn't running.  So there was no
notification that the type wasn't being checked for JS-NATIVEs.

This factors out the code from plain FUNC that does type checking
into an inline-able function routine that fails if a frame's output
cell doesn't match the return type in the spec.