Remove table entries on main thread
The tables which map from the heap pointer of a JS-NATIVE's identity
to a JavaScript function body live on the GUI thread.  (It makes sense
that they do--since the GUI thread is for most purposes the only place
one wants to be running JavaScript code for I/O, DOM manipulation, etc)

However, Rebol code can be run on either the GUI or the worker, and may
trigger a garbage collection.  If on the worker, then a JS-NATIVE which
is no longer needed was being GC'd and triggering an error by running
EM_ASM() code to unregister it.  This was becoming more common with
JS-DO instances that were implemented via JS-NATIVE.

Solving this efficiently will be tricky.  But the recent refactoring to
not lock up the main thread allows an inefficient but workable answer
by using asynchronous posts of the unregister request.  These should
chronologically get in line before any registrations that might happen
against a reused pointer, if the worker chooses to make another
JS-NATIVE immediately after the GC which winds up with the same ID as
the GC'd entity.