Initial support for cancellation
This does some amount of implementation for reb.Halt().  Also replaces
an earlier attempt at trying to tunnel errors through JavaScript via
throws with conventional fail()s, in a way that is less likely to
crash for now.

Cancellation via reb.Halt() would already sort of work (in the pthread
build) if it was Rebol code running, e.g. in a tight loop.  However
there was no way to interrupt a promise that was being awaited on,
e.g. the WAIT command would not be interrupted.  Being able to do so
requires injecting a concept of "cancelability" into promises that
reb.Halt() can call, by being aware of all outstanding promises that
may need to be canceled.

This adds `reb.Cancelable()` as something that can be added to a
promise, as well as automatically adding cancelability to the promises
returned from a JS-AWAITER.