Add #prefer-O2-optimization flag for files in rebmake
Even if a build is trying to optimize for size, it can behoove that
build to favor O2 optimization for certain files...such as %c-eval.c.
This becomes particularly important in the JavaScript build due to
the browser's relatively miserly allowance of how many JS/WASM frames
can be called.  This means stack overflows are hit much more quickly,
so inlining can be of a benefit beyond just performance--by not hitting
the stack overflows.