Hack up JavaScript extension stackoverflow error somewhat
There's a lot of complexity in getting stack overflow errors to work,
as they exist outside the bounds of definition of the C language.
JavaScript adds new issues, by having a separate notion of "stack"
space as used by the WASM heap for C variables from the function call
stack, which is done in terms of JavaScript functions (or WASM calls
that exist on the JavaScript stack).

This makes a small change allowing the JS extension to set a lower
bound for the stack limit, as well as setting it on the promise thread
so it can pick up the thread's limit--vs. the GUI limit.  Since code
doesn't run on both threads at once, this is a workaround that lets
the code run without the previous workaround that didn't check the
stack at all.

In conjunction with the previous commit's allowance of %c-eval.c being
done with -O2 as an override to the -Os, the stack in Chrome seems to
catch the overflow sometimes.  The emterpreter build seems to have
a much lower limit.

A lot more work and thinking needs to be done on this, but it's at
least a slight bit better.