Fix for #prefer-O2-optimization when optimization: false
Not specifying an optimization level led a parent .obj object to use
the "optimization: false" setting which was considered equivalent to
leaving in blank.  (Hence no -O flag on the command line).  This
wasn't picked up on in the code for #prefer-O2-optimization because it
conflated false with blank.

Also, reign => rein.