Treat END of block/group as a valid "eval position"
This enforces going through the evaluator and not "skipping out" on
empty block cases.  e.g. if you write `cycle []` there has to be an
opportunity for Do_Signals_Throws() to check for cancellation
via Ctrl-C.)

This ties into a broader question of considering empty blocks to be
places that are debug step or breakpoint opportunities, so we make
sure you use `do { eval } while (NOT_END(...))` instead of potentially
skipping that opportunity with `while (NOT_END(...)) { eval }`