Primitive Demo of Single-Thread Stepping Concept
(!!! This is a proof-of-concept and very buggy; the debugger extension
should be considered "off limits" at this time.)

While implementing a debugger in a single-thread is not ideal, this
pushes the experimental implementation a slightly further to show how
stepping might be accomplished.

The basic mechanic is that a BREAKPOINT spawns a nested console
session.  This nested console session then signals a RESUME which
causes the session to exit but return a HANDLE!.  That handle is a
hook which is then installed to augment the evaluator.  When its
condition is met, it then unhooks itself so it is no longer monitoring
evaluations, and spawns a console session much as if a BREAKPOINT
had been hit.