Make empty groups or all-comments vaporize
Caution had previously led to an empty group (or group with comments)
to disappear completely in evaluation.  This seemed "too weird", so
it would error:

    >> 1 + () 2
    ** Script error: + is missing its value2 argument

However, the increasing presence of invisibles with multiple arguments
made grouping them potentially desirable.  And a practical case arose
of wanting to be able to put breakpoint-like-things in mid-expression.

This commit changes them to be invisible, within the limits of enfix.
(e.g. `1 () + 2` cannot be invisible).