Remove module isolation from Debugger
Currently the way the ReplPad works is not compatible with module
isolation, because routines like PRINT don't get the updates to the
LIB versions.  As a workaround this removes module isolation from
the debugger.

Also makes the backtrace relative to the debug captured frame, and
not where backtrace is called from.