Move enfixed bit from living in cell to living on action
This mechanically simplifies the notion of the "enfix bit" to live on
the ACTION!, leading to similar behaviors to Rebol2/R3-Alpha/Red
w.r.t. assignment of enfix functions.

This means that to enfix an existing ACTION!, it must be mutated or
it must be copied.  Copying function identities is now a fairly
commonplace operation, so this is less of a big deal than previously.

For the moment, the ENFIX operation is aliased to ENFIXED, which makes
a copy of the function argument.  This is extra cost if enfixing a
function that doesn't need to preserve its non-enfixed form.  However,
review of this point is being saved for considering if enfixedness
is going to be part of the function spec dialect (e.g. "all functions
are enfix, they just may `<skip>` their left argument opportunity).

Also eliminates an unused %r2r3-future.r file, which was renamed to
%bootstrap-shim.r but somehow managed to silently reappear (likely
overlooked during a merge commit).